pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "fatoolsng"
version = "0.1"
description = "DNA fragment analysis tools (next-generation fork)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Alexandr Dorif", email = "dorif11@gmail.com" },
]
keywords = ["dna", "fragment-analysis", "electropherogram", "microsatellite"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "biopython>=1.58",
    "jax",
    "matplotlib",
    "numpy",
    "ruamel.yaml<0.18.0",
    "pandas",
    "scipy",
    "sqlalchemy",
    "transaction",
    "zope.sqlalchemy",
    "sortedcontainers",
    "peakutils",
]

[project.optional-dependencies]
r-analysis = [
    "rpy2",
    "mdp",
]

[project.scripts]
fatoolsng = "fatoolsng.scripts.run:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["fatoolsng/tests"]
python_files = "test_*.py"
python_functions = "test_*"
